Move location resolvers into dedicated modules
